projects/MiroFishTrader/requirements.txt
# MiroFishTrader Dependencies
# Python 3.11+

# Data collection
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
pandas-market-calendars>=4.3.0
pyarrow>=14.0.0

# Technical indicators
ta>=0.11.0

# Reddit
asyncpraw>=7.7.0

# MiroFish client
requests>=2.31.0
httpx>=0.27.0

# LLM (Claude API via litellm proxy)
anthropic>=0.30.0
litellm>=1.40.0

# Slack
# requests already included above

# Env & config
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0
responses>=0.25.0
